// src/app/management/interfaces/pets.interface.ts
// Generated by https://quicktype.io

import { PetPersonality, PetSize, PetStatus,  } from "../constants/pet-enums";

export interface Pet {
    petId:                number;
    petName:              string;
    petAge:               number;
    petSize:              PetSize;
    petPersonality:       PetPersonality;
    petStatus:            PetStatus;
    registrationDateTime: Date;
    rescueDate:           Date;
    rescueHistory:        string;
    petAvatar:            string;
    specialFeatures:      string;
    specieId:             number;
    breedId:              number;
}

export interface Specie {
    specieId:   number;
    specieName: string;
}

export interface Breed {
    breedId:   number;
    breedName: string;
    specie: Specie;
}